isolate component dragging code to a single hook
